libs/state_manager/pyproject.toml
[project]
name = "state-manager"
version = "0.1.4"
description = ""
authors = [
    { name="Tony Truong", email="serinth@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35.71",
    "google-cloud-storage>=2.18.2",
    "jsonpickle>=4.0.0",
    "loguru>=0.7.2",
]

[dependency-groups]
dev = [
    "moto>=5.0.21",
    "pytest>=8.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
